hotkey unit training, extract keyboard logic from ui

+++ game/src/net/protocol/protocol_mod.cpp
@@ -6,7 +6,7 @@ namespace aoe {
 
 void NetPkg::set_protocol(uint16_t version) {
 	PkgWriter out(*this, NetPkgType::set_protocol);
-	write("H", pkgargs{ version }, false);
+	write("H", pkgargs{ (uint64_t)version }, false);
 }
 
 uint16_t NetPkg::protocol_version() {
